+++ node/django_front/src/components/function/Login.jsx
@@ -10,8 +10,8 @@ import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import Container from '@mui/material/Container';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
-import SimpleNavbar from '../SimpleNavbar'
-
+import SimpleNavbar from './SimpleNavbar'
+import NewAccountButton from './NewAccountButton';
 function Copyright(props) {
   return (
     <Typography variant="body2" color="text.secondary" align="center" {...props}>
@@ -27,7 +27,7 @@ function Copyright(props) {
 
 const theme = createTheme();
 
-export default function SignIn() {
+export default function Login() {
   const handleSubmit = (event) => {
     event.preventDefault();
     const data = new FormData(event.currentTarget);
@@ -40,12 +40,13 @@ export default function SignIn() {
 
   return (
     <ThemeProvider theme={theme}>
-      <SimpleNavbar title="ログイン" />
+      <SimpleNavbar/>
       <Container component="main" maxWidth="xs">
+
         <CssBaseline />
         <Box
           sx={{
-            marginTop: 16,
+            marginTop: 8,
             display: 'flex',
             flexDirection: 'column',
             alignItems: 'center',
@@ -86,17 +87,18 @@ export default function SignIn() {
               ログイン
             </Button>
             <Grid container>
-              <Grid item xs>
-                <Link href="#" variant="body2">
-                  アカウントを新規作成
-                </Link>
-              </Grid>
               <Grid item>
                 <Link href="#" variant="body2">
                   {"パスワードを忘れた場合"}
                 </Link>
               </Grid>
             </Grid>
+            <Grid container sx={{ mt: 3}} >
+              <Grid item xs={12} textAlign="center">
+                <Typography>アカウント作成がまだの方はこちらから</Typography>
+              </Grid>
+            </Grid>
+            <NewAccountButton />
           </Box>
         </Box>
         <Copyright sx={{ mt: 16, mb: 4 }} />
