+++ node/django_front/src/components/partTime/PartTimeNavbar.jsx
@@ -1,18 +1,23 @@
 import * as React from 'react';
-import AppBar from '@mui/material/AppBar';
+import { styled, useTheme } from '@mui/material/styles';
 import Box from '@mui/material/Box';
+import MuiAppBar from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
-import Typography from '@mui/material/Typography';
-import NotificationsIcon from '@mui/icons-material/Notifications';
-import Divider from '@mui/material/Divider';
-import Menu from '@mui/material/Menu';
 import List from '@mui/material/List';
-import Badge from '@mui/material/Badge';
+import CssBaseline from '@mui/material/CssBaseline';
+import Divider from '@mui/material/Divider';
 import IconButton from '@mui/material/IconButton';
-import LogoImg from "./title.svg";
-import "../App.css"
-import NotificationListItem from './NotificationListItem';
+import Menu from '@mui/material/Menu';
+import NotificationsIcon from '@mui/icons-material/Notifications';
+import Badge from '@mui/material/Badge';
+import LogoImg from "../titleNavbar.svg";
+import "../../App.css"
+import NotificationListItem from '../NotificationListItem';
+import SettingsIcon from '@mui/icons-material/Settings';
 
+const drawerWidth = 240;
+
+//　連想配列を採用
 const notification = [
   ["店舗登録がまだ終わっていません", "2022/03/03"],
   ["シフト提出の締め切りが迫っています。", "2022/03/21"],
@@ -21,23 +26,54 @@ const notification = [
   ["シフト提出の締め切りが迫っています。", "2022/03/21"],
   ["シフト提出の締め切りが迫っています。", "2022/03/21"],
 ];
-export default function SimpleNavbar() {
+
+const DrawerHeader = styled('div')(({ theme }) => ({
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'flex-end',
+  padding: theme.spacing(0, 1),
+  // necessary for content to be below app bar
+  ...theme.mixins.toolbar,
+}));
+
+const AppBar = styled(MuiAppBar, {
+  shouldForwardProp: (prop) => prop !== 'open',
+})(({ theme, open }) => ({
+  background: "#4DC0B2",
+  zIndex: theme.zIndex.drawer + 1,
+  transition: theme.transitions.create(['width', 'margin'], {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.leavingScreen,
+  }),
+  ...(open && {
+    marginLeft: drawerWidth,
+    width: `calc(100% - ${drawerWidth}px)`,
+    transition: theme.transitions.create(['width', 'margin'], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+  }),
+}));
+
+export default function PartTimeNavbar(props) {
+  const theme = useTheme();
+  const [open, setOpen] = React.useState(false);
   const [anchorElNotification, setAnchorElNotification] = React.useState(null);
+
   const handleOpenNotificationMenu = (event) => {
     setAnchorElNotification(event.currentTarget);
   };
-  
+
   const handleCloseNotificationMenu = () => {
     setAnchorElNotification(null);
   };
 
   return (
-    <Box sx={{ flexGrow: 1 }}>
-      <AppBar position="static" sx={{
-          background: "#4DC0B2"
-      }}>
+    <Box sx={{ display: 'flex' }}>
+      <CssBaseline />
+      <AppBar position="fixed" open={open}>
         <Toolbar>
-        <a href='localhost:3000' className='App-logo'><img src={LogoImg} alt="logo" className='App-logo'></img></a>
+          <a href='localhost:3000' className='App-logo'><img src={LogoImg} alt="logo" className='App-logo'></img></a>
           <Box sx={{ flexGrow: 1 }} />
           <Box>
             <IconButton
@@ -49,6 +85,13 @@ export default function SimpleNavbar() {
               <Badge badgeContent={17} color="error">
                 <NotificationsIcon />
               </Badge>
+            </IconButton>
+            <IconButton
+              size="large"
+              aria-label="show 17 new notifications"
+              color="inherit"
+            >
+              <SettingsIcon />
             </IconButton>
             <Menu
               sx={{ mt: '45px' }}
@@ -70,7 +113,7 @@ export default function SimpleNavbar() {
               <List>
                 {notification.map((text) => (
                   <>
-                    <NotificationListItem primaryText={text[0]} secondaryText={text[1]}/>
+                    <NotificationListItem primaryText={text[0]} secondaryText={text[1]} />
                     <Divider />
                   </>
                 ))}
@@ -79,6 +122,10 @@ export default function SimpleNavbar() {
           </Box>
         </Toolbar>
       </AppBar>
+      <Box component="main" sx={{ flexGrow: 1, p: 3 }}>
+        <DrawerHeader />
+        {props.contents}
+      </Box>
     </Box>
   );
 }
