+++ node/django_front/src/components/store/Navbar.jsx
@@ -24,9 +24,9 @@ import EventNoteOutlinedIcon from '@mui/icons-material/EventNoteOutlined';
 import DriveFileRenameOutlineOutlinedIcon from '@mui/icons-material/DriveFileRenameOutlineOutlined';
 import NotificationsIcon from '@mui/icons-material/Notifications';
 import Badge from '@mui/material/Badge';
-import LogoImg from "./titleNavbar.svg";
-import "../App.css"
-import NotificationListItem from './NotificationListItem';
+import LogoImg from "../titleNavbar.svg";
+import "../../App.css"
+import NotificationListItem from '../NotificationListItem';
 
 const drawerWidth = 240;
 
@@ -106,7 +106,7 @@ const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' 
   }),
 );
 
-export default function MiniDrawer(props) {
+export default function Navbar(props) {
   const theme = useTheme();
   const [open, setOpen] = React.useState(false);
   const [anchorElNotification, setAnchorElNotification] = React.useState(null);
