+++ node/django_front/src/components/store/Register.jsx
@@ -11,7 +11,7 @@ import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import Container from '@mui/material/Container';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
-import SimpleNavbar from '../SimpleNavbar';
+import SimpleNavbar from '../function/SimpleNavbar';
 import lightGreen from '@mui/material/colors/lightGreen';
 
 function Copyright(props) {
@@ -33,7 +33,7 @@ const theme = createTheme({
   },
 });
 
-export default function SignUp() {
+export default function Register() {
   const handleSubmit = (event) => {
     event.preventDefault();
     const data = new FormData(event.currentTarget);
@@ -148,7 +148,7 @@ export default function SignUp() {
                 店舗情報登録
                 </Typography>
               </Grid>
-              <Grid item xs={12} sm={6}>
+              <Grid item xs={12} >
                 <TextField
                   required
                   fullWidth
